+++ source/G3Dcpp/Vector3int16.cpp
@@ -16,9 +16,9 @@
 namespace G3D {
 
 Vector3int16::Vector3int16(const class Vector3& v) {
-    x = iFloor(v.x + 0.5);
-    y = iFloor(v.y + 0.5);
-    z = iFloor(v.z + 0.5);
+    x = (int16)iFloor(v.x + 0.5);
+    y = (int16)iFloor(v.y + 0.5);
+    z = (int16)iFloor(v.z + 0.5);
 }
